Random Password Generator
This is day 5 of the 100 days of coding in Python. Today we were tasked with creating a program that will take the number of desired letters, symbols, and numbers. It will then randomize a password for you. There are two versions: easy (where the letters, numbers, and symbols are not mixed) and hard (where they are).